Add font loading and rendering
This also adds path blocked indicator to the center of mouse hex. Note
that it's not currently possible to replicate the original style where
the text inside of hex appears to have thick outline. This is the effect
of drawing outline twice and clipping to 1 px rect around the original
glyph texture.